+++ server/component/style/validate/ValidateController.php
@@ -44,6 +44,10 @@ class ValidateController extends BaseController
                 if (isset($_POST['mobile']) && $_POST['mobile']) {
                     $this->success_msgs[] = $this->model->get_db_field('alert_success');
                 }
+                if ($this->model->get_redirect_page_keyword()) {
+                    $this->model->get_services()->get_login()->check_credentials($this->model->get_user_email(), $_POST['pw']);
+                    header('Location: ' . $this->model->get_link_url($this->model->get_redirect_page_keyword()));
+                }
             } else {
                 $this->fail = true;
                 if (isset($_POST['mobile']) && $_POST['mobile']) {